requirements.txt
Django==5.0.2
factory-boy==3.3.0